Recover ResourceEntry as a real class; wire ResourceManager::Init()
ResourceManager::GetById was already confirmed correct; the crash chain
this session (and the previous one) traced back to
ResourceManager::Init() ultimately hit RESMGR_AllocResourceEntry, a
completely unimplemented void-returning stub, and a genuine type-safety
bug independent of that: destroy_resource/create_string_resource
cross-cast every handle to ResourceObject* before calling delete, even
though ResourceEntry is a wholly separate, unrelated class hierarchy
(different vtable, different slot count).

ResourceEntry is now a real polymorphic class with:
- two real constructors (0x448990 sound-resource form, 0x448A20
  external-file form)
- a real virtual destructor (0x4489D0)
- a private OpenResourceFile() (0x448A70) using a real WIN32_Stream RAII
  local, matching game/TrainStation.cpp's already-converted sibling
  idiom, with the archive/disk/GetFileAttributesA fallback chain
  recovered via disassembly (not decompiler pseudocode alone -- the
  pseudocode's own variable-splitting misrepresented which buffer the
  .dat-extension swap mutated)
- a real virtual Parse() (0x448C90 -- had no Ghidra function boundary
  at all before this session), a line-oriented key=value descriptor
  parser, backed by new real WNDPROC_Stream::ExtractInt/
  ExtractUnsignedInt/ReadNumericToken (0x4646C0/0x464F70/0x465AD0) and
  WNDPROC_StreamBuf::PutBack (0x4657A0) implementations

destroy_resource is split into destroy_window_resource (ChildWindow*,
for the resource_ptrs registry) and destroy_resource_entry
(ResourceEntry*, for the string_cache registry) -- ResourceManager::
FreeAllResources already treats these as two separate arrays, so the
split reflects real, pre-existing structure rather than a new
distinction.

Parse/ReadNumericToken/PutBack are confirmed unreachable on this host
(g_asset_mgr.archive_file is always 0 in retail; the host
GetFileAttributesA stub always reports "missing", forcing is_valid=1
before ever opening a real stream) -- verified via temporary [REACH]
instrumentation across the full integration suite before removing it.
Implemented from decompilation and reviewed, but not exercised.

Wiring g_resmgr.Init() into GameLoop_Setup surfaced two more crashes in
already-INTEGRATED code that had simply never executed before (AddString's
resource sweep never ran until Init() did): input/TrackTileDescriptor.cpp
and game/TrainStation.cpp both read an `extern` global annotated with the
original PE address 0x479190 but never actually defined on this host,
silently linked to a null GOT slot under
-Wl,--unresolved-symbols=ignore-all. Both fixed by using the real
Win32Stream.h kStreamShareMask constant that address holds. Also fixed a
duplicate-alias landmine found in the same file: TrackTileDescriptor.cpp's
"g_scene_name" was a separately-defined, permanently-empty duplicate of
g_install_path.

The game now boots through Init(), InitAllSubsystems, DDRAW_Init, the
launch intro, and into the main menu/mode-2 loop with no crash. The GUI
integration suite returns to the exact pre-existing baseline (10
passed/5 failed, identical failing test names) with zero regression,
verified via git-stash A/B plus a byte-identical 54/54 mingw typecheck
failure set. The 5 remaining failures are the original, pre-existing
PostcardAlbum::InitWindowSurface crash (re-confirmed via coredump,
identical stack trace) -- not a regression, and not yet fixed; see
PROGRESS.md for the sharper lead this session found (a ChildWindow-vs-
ResourceObject hierarchy mismatch analogous to the one just fixed here).

+++ shared/stubs_link001_batch3_resource_audio.cpp
@@ -294,12 +294,13 @@ void ReleaseSoundResource(int handle)
 /* ReleaseSoundResource(ResourceEntry*)                                 */
 /* Address: 0x448EE0                                                   */
 /*                                                                     */
-/* Declared in resources/ResourceManager.h:828. Decompiled in full      */
+/* Declared in resources/ResourceManager.h. Decompiled in full           */
 /* (2026-08-15): decrements refcount at +0x120 (never below 0); when it */
-/* reaches 0 and a DirectSound buffer exists at +0x0C, checks the exact  */
-/* low byte at +0x08 against the literal 1 (`*(char*)(this+8) != 1`,     */
-/* NOT a bitmask test — flags is an int16_t, so this reads only its low  */
-/* byte) before stopping (vtable slot 0x48/4=18) and releasing (vtable   */
+/* reaches 0 and a DirectSound buffer exists at +0x0C, checks `flags`    */
+/* (+0x08, a real `uint8_t` — confirmed via raw disassembly of           */
+/* ResourceEntry::OpenResourceFile's single-byte zeroing write, 2026-08- */
+/* 21) against the literal 1 (`*(char*)(this+8) != 1`, an exact-byte-    */
+/* equals test, NOT a bitmask) before stopping (vtable slot 0x48/4=18) and releasing (vtable   */
 /* slot 8/4=2) the buffer and clearing the field. `entry->buffer` is now */
 /* typed `AudioDirectSoundBuffer*` (resources/ResourceManager.h), the    */
 /* same COM interface audio/AudioChannel.cpp already uses for its own    */
@@ -316,7 +317,7 @@ int32_t ReleaseSoundResource(ResourceEntry* entry)
         entry->refcount--;
     }
     if (entry->refcount <= 0 && entry->buffer != nullptr &&
-        static_cast<uint8_t>(entry->flags) != 1) {
+        entry->flags != 1) {
         entry->buffer->Stop();
         entry->buffer->Release();
         entry->buffer = nullptr;
